+++ src/utils/intrinsic-functions-schema.js
@@ -1,52 +1,62 @@
 const yaml = require('js-yaml');
 
-module.exports = yaml.Schema.create([
-	new yaml.Type('!Base64', {
-		kind: 'scalar',
-		construct: data => ({ 'Fn::Base64': data }),
-	}),
-	new yaml.Type('!Cidr', {
-		kind: 'sequence',
-		construct: data => ({ 'Fn::Cidr': data }),
-	}),
-	new yaml.Type('!FindInMap', {
-		kind: 'sequence',
-		construct: data => ({ 'Fn::FindInMap': data }),
-	}),
-	new yaml.Type('!GetAtt', {
-		kind: 'scalar',
-		construct: data => ({ 'Fn::GetAtt': data }),
-	}),
-	new yaml.Type('!GetAZs', {
-		kind: 'scalar',
-		construct: data => ({ 'Fn::GetAZs': data }),
-	}),
-	new yaml.Type('!ImportValue', {
-		kind: 'scalar',
-		construct: data => ({ 'Fn::ImportValue': data }),
-	}),
-	new yaml.Type('!Join', {
-		kind: 'sequence',
-		construct: data => ({ 'Fn::Join': data }),
-	}),
-	new yaml.Type('!Select', {
-		kind: 'sequence',
-		construct: data => ({ 'Fn::Select': data }),
-	}),
-	new yaml.Type('!Split', {
-		kind: 'scalar',
-		construct: data => ({ 'Fn::Split': data }),
-	}),
-	new yaml.Type('!Sub', {
-		kind: 'scalar',
-		construct: data => ({ 'Fn::Sub': data }),
-	}),
-	new yaml.Type('!Transform', {
-		kind: 'mapping',
-		construct: data => ({ 'Fn::Transform': data }),
-	}),
-	new yaml.Type('!Ref', {
-		kind: 'scalar',
-		construct: data => ({ Ref: data }),
-	}),
-]);
+const types = [];
+
+const specialTypes = {
+	Ref: 'Ref',
+	Condition: 'Condition',
+};
+
+const schema = {
+	scalar: [
+		'Base64',
+		'GetAtt',
+		'GetAZs',
+		'ImportValue',
+		'Sub',
+		'Ref',
+		'Condition',
+	],
+	sequence: [
+		'Cidr',
+		'FindInMap',
+		'Join',
+		'Select',
+		'And',
+		'Equals',
+		'If',
+		'Not',
+		'Or',
+		'Split',
+	],
+	mapping: [
+		'Transform',
+	],
+};
+
+const constructs = {
+	GetAtt(data) {
+		const parts = data.split('.');
+		return {
+			'Fn::GetAtt': [
+				parts[0],
+				parts.slice(1).join('.'),
+			],
+		};
+	},
+};
+
+Object.keys(schema).forEach((kind) => {
+	schema[kind].forEach((name) => {
+		const fn = specialTypes[name] || `Fn::${name}`;
+		const params = {
+			kind,
+			construct: constructs[name] || ((data) => ({ [fn]: data })),
+		};
+
+		const type = new yaml.Type(`!${name}`, params);
+		types.push(type);
+	});
+});
+
+module.exports = yaml.Schema.create(types);