+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Updated dependencies to fix vulnerability issues found in dependant packages.
 - Added bash/zsh-completion shortcuts for commands.
+- Added sample project that demonstrates how to use cfpack to deploy Appsync + Lambda infrastructure.
 
 ## v1.2.1 (2019-06-14)